backend/requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
websockets
scikit-learn
sqlalchemy 
psycopg2-binary
